update: lazy commit message, formatting and fixing code styling, going to introduce state
