+++ src/app/shared/components/entity-hider/entity-hider.component.html
@@ -1,14 +1,18 @@
 <div class="rd-grouping mt-2">
-  <div class="flex flex-col justify-center space-y-8  h-full">
-    <div class="descriptor flex flex-col space-y-2">
-        <h1 class="text-white font-semibold text-xl">Entity Hider</h1>
-        <p>Almost exclusively for W79 Divination.</p>
-    </div>
+    <div class="flex flex-col justify-center space-y-8  h-full">
+        <div class="descriptor flex flex-col space-y-2">
+            <h1 class="text-white font-semibold text-xl">Entity Hider</h1>
+            <p>Almost exclusively for W79 Divination.</p>
+        </div>
 
-    <div class="flex justify-between">
-        <div><p-button (click)="hideEntities()" class="w-56">Entities</p-button></div>
-        <div><p-button (click)="hideNpc()" class="w-56">Npcs</p-button></div>
+        <div class="flex justify-between">
+            <div>
+                <p-button (click)="hideEntities()" class="w-56">Entities</p-button>
+            </div>
+            <div>
+                <p-button (click)="hideNpc()" class="w-56">Npcs</p-button>
+            </div>
+        </div>
     </div>
-  </div>
 </div>
