+++ src/app/shared/components/entity-hider/entity-hider.component.ts
@@ -1,7 +1,7 @@
-import { Component } from '@angular/core';
-import { queuePacket } from '../../../icp-events/events';
-import { FormsModule } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
+import {Component} from '@angular/core';
+import {queuePacket} from '../../../icp-events/events';
+import {FormsModule} from '@angular/forms';
+import {ButtonModule} from 'primeng/button';
 
 @Component({
   selector: 'app-entity-hider',
@@ -11,8 +11,6 @@ import { ButtonModule } from 'primeng/button';
   styleUrl: './entity-hider.component.css'
 })
 export class EntityHiderComponent {
-
-
   hideEntities() {
     queuePacket("req", "hideentities", "entity");
   }
